Exceptions Handlers Ajuste No Controller etc

+++ src/main/java/com/movierental/service/UserService.java
@@ -1,5 +1,6 @@
 package com.movierental.service;
 
+import com.movierental.exception.BusinessException;
 import com.movierental.model.User;
 import com.movierental.repository.UserMovieRepository;
 import com.movierental.repository.UserRepository;
@@ -8,6 +9,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.sql.DataSource;
+import java.util.Objects;
 import java.util.regex.Pattern;
 
 @Service
@@ -24,19 +26,26 @@ public class UserService {
         this.userMovieRepository = userMovieRepository;
     }
 
-    public String insertUser(User newUser){
+    public String insertUser(User newUser) throws BusinessException {
+        if(Objects.isNull(newUser)){
+            throw new BusinessException("Objeto User do parametro newUser está null");
+        }
         User user = userRepository.findByUsername(newUser.getUsername());
         if(user != null){
             return "O email " + newUser.getUsername() + " já está cadastrado. Inserir outro!";
         }else{
             boolean validName = validateName(newUser.getName());
             boolean validEmail = validateEmail(newUser.getUsername());
+            boolean validPassword = validatePassword(newUser.getPassword());
 
             if(!validName){
                 return "O nome está vazio ou inválido.";
             }
             if (!validEmail){
                 return "Seu email não é válido";
+            }
+            if(!validPassword){
+                return "O password está irregular. Deve ser menos de 32 caracteres ou está vazio";
             }
             newUser.setPassword(generatePassword(newUser.getPassword()));
             newUser.setRole("ROLE_USER");
@@ -48,7 +57,10 @@ public class UserService {
         }
     }
 
-    public String deleteUser(String userName){
+    public String deleteUser(String userName) throws BusinessException {
+        if(Objects.isNull(userName)){
+            throw new BusinessException("Username is null, shouldn't be");
+        }
         User user = userRepository.findByUsername(userName);
         userRepository.delete(user);
         return "O usuário: " + user.getUsername() + " foi deletado do sistema.";
@@ -73,6 +85,13 @@ public class UserService {
     private boolean validateEmail(String email) {
         final Pattern EMAIL_REGEX = Pattern.compile("[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?", Pattern.CASE_INSENSITIVE);
         return EMAIL_REGEX.matcher(email).matches();
+    }
+
+    private boolean validatePassword(String password){
+        if(Objects.isNull(password) || password.isEmpty() || password.length() > 32){
+            return false;
+        }
+        return true;
     }
